add cer window

+++ src/main/java/com/core/service/repositories/RepositoryAgencias.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RepositoryAgencias extends CrudRepository<Agencias,Long> {
+public interface RepositoryAgencias extends CrudRepository<Agencias,Integer> {
 }
